new dataformat for amplitudes

+++ test/TrackLocalAngle.cc
@@ -53,7 +53,7 @@ std::pair<float,float> TrackLocalAngle::findhitcharge(const TrajectoryMeasuremen
     // THIS IS THE POINTER TO THE MONO HIT OF A MATCHED HIT 
     const SiStripRecHit2D *monohit=matchedhit->monoHit();    
     const SiStripCluster* monocluster = &*(monohit->cluster());
-    const std::vector<uint16_t> amplitudesmono( monocluster->amplitudes().begin(),
+    const std::vector<uint8_t> amplitudesmono( monocluster->amplitudes().begin(),
 						monocluster->amplitudes().end());
     for(size_t ia=0; ia<amplitudesmono.size();ia++)
       {
@@ -63,7 +63,7 @@ std::pair<float,float> TrackLocalAngle::findhitcharge(const TrajectoryMeasuremen
     // THIS IS THE POINTER TO THE STEREO HIT OF A MATCHED HIT 
     const SiStripRecHit2D *stereohit=matchedhit->stereoHit();   
     const SiStripCluster* stereocluster = &*(stereohit->cluster());
-    const std::vector<uint16_t> amplitudesstereo( stereocluster->amplitudes().begin(),
+    const std::vector<uint8_t> amplitudesstereo( stereocluster->amplitudes().begin(),
 						  stereocluster->amplitudes().end());
     for(size_t ia=0; ia<amplitudesstereo.size();ia++)
       {
@@ -75,7 +75,7 @@ std::pair<float,float> TrackLocalAngle::findhitcharge(const TrajectoryMeasuremen
     
     //  hit= POINTER TO THE RECHIT
     const SiStripCluster* cluster = &*(hit->cluster());
-    const std::vector<uint16_t> amplitudes( cluster->amplitudes().begin(),
+    const std::vector<uint8_t> amplitudes( cluster->amplitudes().begin(),
 					    cluster->amplitudes().end());
     for(size_t ia=0; ia<amplitudes.size();ia++)
       {
